defuzzification

+++ src/com/App.java
@@ -5,9 +5,9 @@ public class App {
     public static void main(String[] args) throws Exception {
         Scanner input = new Scanner(System.in);
         System.out.println("Enter Project Fund : ");
-        int projectFund = input.nextInt();
+        Double projectFund = input.nextDouble();
         System.out.println("Enter Experience Level : ");
-        int ExperienceLevel = input.nextInt();
+        Double ExperienceLevel = input.nextDouble();
         FuzzyLogic testCase = new FuzzyLogic(projectFund, ExperienceLevel);
         testCase.performFL();
         input.close();
