+++ src/com/tuple.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 
 public class tuple{
     String name;
-    ArrayList<Integer> set = new ArrayList<Integer>();
+    ArrayList<Double> set = new ArrayList<Double>();
     ArrayList<Double> output = new ArrayList<Double>();
 
-    public tuple(String name, ArrayList<Integer> set, ArrayList<Double> output)
+    public tuple(String name, ArrayList<Double> set, ArrayList<Double> output)
     {
         this.name = name;
         this.set = set;
